pyproject.toml
[project]
name = "zenstock"
version = "0.1.0"
description = "ZenStock: a stock manager (to keep zen 😊)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-environ>=0.12.0",
    "django>=5.2",
    "pytest-django>=4.11.1",
    "django-crispy-forms>=2.4",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "zenstock.settings.dev"
python_files = ["test_*.py",]
